`MockFactory` <--> `MockRepository`
`MockFactory` has long been deprecated but still holds all the members
that ought to be in `MockRepository`. This commit inverts the inherit-
ance relationship between the two classes so that we can move over the
members into today's main class.
